Fix Bootstrap 2.x incompatibility with jQuery 3.x
Bootstrap 2.3.2 has an incomaptibility with jQuery 3.x causing dropdown
menus to break.

Unfortunately the Bootstrap 2.x series is not maintained anymore and
migrating to Bootstrap 3 or 4 is quite expensive.

Therefore we only apply a quickfix by removing the hash from the
following expression in the bootstrap source of the Dropdown and Alert:

```
.replace(/.*(?=#[^\s]*$)/,""))
```

In case of bootstrap tabs we can simply remove the entire line.

See also:
https://github.com/FezVrasta/bootstrap-material-design/issues/1137

https://machtfit.atlassian.net/browse/MAC-2194
https://machtfit.atlassian.net/browse/MAC-2276